requirements.txt
django==5.1.4
djangorestframework==3.16.0
wagtail==6.4.1
djangorestframework-simplejwt==5.5.0
